Fix StyleCop warnings in BaseNumberRangeExtractor - Extract NumberRangeContants - Suppress stylecop warning in constants variable names
